Fix datasource issue
Currently the playbook can only run once and will fail once the datasource exists. This means we cannot run the job to carry out updates to dashboards etc as the playbook is bombing out.

The fix essentially looks to see if the datasource specified exists and if it does skip the create.

Partially resolves: DVOP-2303