Add kafka overview dashboard
As part of the kafka exporter poc a while back, I found a dashboard that would provide an overview of all the topics within a kafka cluster.
FESS has found this dashboard very useful and therefore I agreed to get this added to our provisioning.

Adds a templated dashboard that is usable across all environments and adds a task so its provisioned out of the box.

Resolves: DVOP-2312